Don't respond with full details in case of HTP Endpoint faults for sec
